Massive overhaul of all Perl code
Modernized and standardized:

- Pragmas
- Usage of signatures
- 'use parent' for inheritance
- All invocations of 'new' changed to arrow notation ('create', etc.
  changed to 'new')
- Changed @EXPORT -> @EXPORT_OK. All OO libraries export nothing.
- Where possible, use 'use' to include modules at compile time with
  minimal imports.
- Whitespace
- Fixed deprecated (or disfavoured) patterns, like bareword variables
  and 2 argument opens.
- Changed $self -> $this
- Changed function names using built-in keywords.
- All possible function names made snake_case.
- Passing Perl::Critic level 4 (with minor exceptions; test and
  config files added)
- Changed all local incarnations of 'ReadConfig' to import the
  STUtils.pm version instead.

Also:

- Removed some Enterprise-only/registration code. Flagged others where
  they might be able to have subtitute workflows.
- Changed how watchdogs will fetch custom scritps (since all of
  /usr/spamtagger will be read-only).
- Moved some files to better-suited locations.

+++ bin/watchdog/ST_mod_SKELETON.pl
@@ -23,13 +23,8 @@ open my $file, '>', $OUT_FILE;
 # Fonction de gestion de la sortie du script
 # A appeler également en cas de succès
 # Efface le fichier PID et renseigne le temps d'éxécution et le return code dans le fichier de sortie
-sub my_own_exit {
-    my ($exit_code) = @_;
-    $exit_code = 0  if ( ! defined ($exit_code) );
-
-    if ( -e $PID_FILE ) {
-        unlink $PID_FILE;
-    }
+sub my_own_exit ($exit_code = 0) {
+    unlink $PID_FILE if ( -e $PID_FILE );
 
     my $ELAPSED = time() - $timestamp;
     print $file "EXEC : $ELAPSED\n";
@@ -40,13 +35,8 @@ sub my_own_exit {
     exit($exit_code);
 }
 
-
 #### MAIN
 #### Lorsque le module a trouvé une erreur, il est censé sortir avec my_own_exit "#ERREUR" (avec #ERREUR : chiffre : retour de la commande)
-
-
-
-
 
 # A CONSERVER !
 my_own_exit(0);
