+++ bin/watchdog/ST_mod_detect_bad_DKIM.pl
@@ -18,61 +18,46 @@ my $OUT_FILE   = '/var/spamtagger/spool/watchdog/' .$script_name_no_ext. '_' .$t
 
 open my $file, '>', $OUT_FILE;
 
-sub my_own_exit {
-    my ($exit_code) = @_;
-    $exit_code = 0  if ( ! defined ($exit_code) );
+sub my_own_exit ($exit_code = 0) {
+  unlink $PID_FILE if ( -e $PID_FILE );
 
-    if ( -e $PID_FILE ) {
-        unlink $PID_FILE;
-    }
+  my $ELAPSED = time() - $timestamp;
+  print $file "EXEC : $ELAPSED\n";
+  print $file "RC : $exit_code\n";
 
-    my $ELAPSED = time() - $timestamp;
-    print $file "EXEC : $ELAPSED\n";
-    print $file "RC : $exit_code\n";
+  close $file;
 
-    close $file;
-
-    exit($exit_code);
+  exit($exit_code);
 }
 
 opendir (my $dir, '/var/spamtagger/spool/tmp/spamtagger/dkim/');
 my @short;
 my @invalid;
 while (my $key = readdir($dir)) {
-    if ($key eq 'default.pkey' && -s '/var/spamtagger/spool/tmp/spamtagger/dkim/'.$key <= 1) {
-        next;
-    }
-    if ($key =~ m/^\.+$/) {
-        next;
-    }
-    my $length = `openssl rsa -in /var/spamtagger/spool/tmp/spamtagger/dkim/$key -noout -text 2> /dev/null | grep 'Private-Key:'` || 'invalid';
-    chomp($length);
-    $length =~ s/Private-Key: \((\d+) bit\)/$1/;
-    if ($length =~ m/^\d+$/) {
-        if ($length < $min_length) {
-            push(@short, $key);
-        }
-    } else {
-        push(@invalid, $key);
-    }
+  next if ($key eq 'default.pkey' && -s '/var/spamtagger/spool/tmp/spamtagger/dkim/'.$key <= 1);
+  next if ($key =~ m/^\.+$/);
+  my $length = `openssl rsa -in /var/spamtagger/spool/tmp/spamtagger/dkim/$key -noout -text 2> /dev/null | grep 'Private-Key:'` || 'invalid';
+  chomp($length);
+  $length =~ s/Private-Key: \((\d+) bit\)/$1/;
+  if ($length =~ m/^\d+$/) {
+    push(@short, $key) if ($length < $min_length);
+  } else {
+    push(@invalid, $key);
+  }
 }
 
 my $status = '';
 my $rc = 0;
 if (scalar(@short)) {
-    $rc += 1;
-    $status .= 'Short DKIM key length: ' . join(', ', @short);
+  $rc += 1;
+  $status .= 'Short DKIM key length: ' . join(', ', @short);
 }
 if (scalar(@invalid)) {
-    if ($rc) {
-        $status .= '<br/>';
-    }
-    $rc += 2;
-    $status .= 'Invalid DKIM key: ' . join(', ', @invalid);
+  $status .= '<br/>' if ($rc);
+  $rc += 2;
+  $status .= 'Invalid DKIM key: ' . join(', ', @invalid);
 }
 
-if ($status) {
-    print $file $status."\n";
-}
+print $file $status."\n" if ($status);
 
 my_own_exit($rc);
